requirements.txt
gin-config
seqio
pytorch-crf
transformers
sentencepiece
datasets==1.5.0